requirements.txt
mcp
chainlit
google-generativeai
groq
streamlit
groq